code/my-react-app/hf-space-whisper/requirements.txt
gradio>=4.44.0,<5.0.0
openai-whisper==20240930
torch>=2.2.0,<3.0.0
numpy>=1.26.0,<3.0.0
scipy>=1.11.0,<2.0.0
fastapi>=0.110.0,<1.0.0
uvicorn>=0.29.0,<1.0.0
